openAI/telegram_chat_bot/requirements.txt
aiogram
openai==0.28
python-dotenv

